conference: clear participant data on participant removal
Currently removing a particpant from a conference removes the subcall from the conference, but does not handle the clearing of the leaving participant's data in the conference. This is problematic as it means the participant is still alive in the conference infos, causing stale data and an outdated list of participants until a new participant is added where the list is properly rebuilt.

This patch:
* adds the clearParticipantData(const std::string& callId) method to the Conference class
* adds a call to clearParticipantData inside removeSubCall right after removing the subcall from the conference

GitLab: #1159
Change-Id: I454bab8bb34926097120fe53d5ccb76e0e78cd3e